Move docs types
